Separating CMS schema/yup validation so we can use yup on front-end
The export of mongoose.model was breaking the front-end as the .model is not available on the front end
